feat(codegen): codegen -> JavaScriptCodegen, codegen dict
