+++ packages/codegen/src/codegens/index.ts
@@ -1,0 +1,8 @@
+import { JavaScriptCodegen } from './JavaScriptCodegen.js';
+
+export const codegens = {
+  JavaScript: JavaScriptCodegen,
+  TypeScript: JavaScriptCodegen,
+};
+
+export default codegens;
