+++ packages/codegen/src/types.ts
@@ -1,3 +1,5 @@
+import codegens from './codegens/index.js';
+
 export interface ICodegenOptions {
   className: string;
 }
@@ -12,3 +14,12 @@ export interface ITypedObjectProps {
 export interface ITypedObject extends ITypedObjectProps {
   codeBlock: string;
 }
+
+export interface ICodegen {
+  dictionaries: string[];
+  changeStateHandlers: string[];
+  handlersDict: string[];
+  getClassTemplate(className: string): string;
+}
+
+export type TCodegenType = keyof typeof codegens;
